+++ src/main/java/com/tm30/structmenu/repository/StrategyRepository.java
@@ -3,10 +3,25 @@ package com.tm30.structmenu.repository;
 import com.tm30.structmenu.strategy.Strategy;
 
 import java.io.*;
+import java.util.Optional;
 
 public class StrategyRepository {
 
     static public void save(Strategy strategy) {
+
+        try {
+            ObjectOutputStream oos = new ObjectOutputStream(
+                    new FileOutputStream(new File("/tmp/" + strategy.getSlug().replace("//s+", "_").toLowerCase()))
+            );
+
+            oos.writeObject(strategy);
+
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    static public void saveById(Strategy strategy) {
 
         try {
             ObjectOutputStream oos = new ObjectOutputStream(
@@ -19,18 +34,16 @@ public class StrategyRepository {
         }
     }
 
-    static public boolean exists(String id) {
-
-        File file = new File("/tmp/" + id);
-        return file.exists();
-
+    static public boolean exists(String strategySlug) {
+        File fileName = new File("/tmp/" + strategySlug);
+        return fileName.exists();
     }
 
-    static public Strategy getStrategy(String id) {
+    static public Strategy getStrategy(String strategySlug) {
         Strategy strategy = null;
 
         try {
-            File file = new File("/tmp/" + id);
+            File file = new File("/tmp/" + strategySlug);
 
             if (file.exists()) {
                 ObjectInputStream oos = new ObjectInputStream(
@@ -49,6 +62,8 @@ public class StrategyRepository {
 
     static public Boolean delete(Strategy strategy) {
 
+        if (Optional.ofNullable(strategy).isPresent()) return false;
+
         File f = new File("/tmp/" + strategy.getUniqueId());
 
         return f.delete();
@@ -56,4 +71,25 @@ public class StrategyRepository {
     }
 
 
+    public static Strategy getStrategyById(String strategyId) {
+
+        Strategy strategy = null;
+
+        try {
+            File file = new File("/tmp/" + strategyId);
+
+            if (file.exists()) {
+                ObjectInputStream oos = new ObjectInputStream(
+                        new FileInputStream(file));
+
+                strategy = (Strategy) oos.readObject();
+            }
+
+        } catch (ClassNotFoundException | IOException e) {
+            e.printStackTrace();
+        }
+
+        return strategy;
+
+    }
 }
